Added a router wrapper for "head" requests.
This shouldn't really be used on the client, but I have a shared
(client/server) router configuration that expected this function to
exist, even though it wouldn't be called on the client.

Feels ugly putting this in, but for now at least, it makes my life a
little easier to do it this way.